Remove JSONLoader from core
