+++ docs/api/zh/loaders/ObjectLoader.html
@@ -13,8 +13,6 @@
 		<p class="desc">
 			A loader for loading a JSON resource in the [link:https://github.com/mrdoob/three.js/wiki/JSON-Object-Scene-format-4 JSON Object/Scene format].<br /><br />
 
-			Note that this loader can't load geometries of type [page:Geometry]. Use [page:JSONLoader] instead for that.<br /><br />
-
 			此加载器内部使用[page:FileLoader]进行加载文件。
 		</p>
 
@@ -118,7 +116,6 @@
 		[page:Object json] — 必选参数，需要被解析的JSON源。<br /><br />
 
             此函数以JSON结构，用[page:.parse]去解析[page:Geometry geometries]或[page:BufferGeometry buffer geometries]。
-            在内部，它使用JSONLoader作为几何加载器，使用BufferGeometryLoader作为几何缓冲区加载器。
 		</p>
 
 		<h3>[method:Object3D parseMaterials]( [param:Object json] )</h3>
