+++ docs/manual/zh/introduction/Animation-system.html
@@ -32,7 +32,7 @@
 			<br /><br />
 			每个* AnimationClip *通常保存对象的某个活动的数据。 如果
 			mesh是一个字符，例如，可以有一个用于walkcycle的动画片段，第二个
-			跳跃，三分之一的回避等等。		
+			跳跃，三分之一的回避等等。
 		</p>
 
 		<h3>Keyframe Tracks(关键帧轨道)</h3>
@@ -83,11 +83,10 @@
 		<p class="desc">
 			请注意，并非所有模型格式都包含动画（尤其是OBJ，没有），而且只有一些
 			three.js加载器支持[page：AnimationClip AnimationClip]序列。 以下几个<i>确实</ i>
-			支持此动画类型：		
+			支持此动画类型：
 		</p>
 
 			<ul>
-				<li>[page:JSONLoader THREE.JSONLoader]</li>
 				<li>[page:ObjectLoader THREE.ObjectLoader]</li>
 				<li>THREE.BVHLoader</li>
 				<li>THREE.ColladaLoader</li>
