+++ test/three.source.unit.js
@@ -161,7 +161,6 @@ import './unit/src/loaders/DataTextureLoader.tests';
 import './unit/src/loaders/FileLoader.tests';
 import './unit/src/loaders/FontLoader.tests';
 import './unit/src/loaders/ImageLoader.tests';
-import './unit/src/loaders/JSONLoader.tests';
 import './unit/src/loaders/Loader.tests';
 import './unit/src/loaders/LoaderUtils.tests';
 import './unit/src/loaders/LoadingManager.tests';
