+++ test/unit/src/geometries/ConeGeometry.tests.js
@@ -5,45 +5,10 @@
 
 import { runStdGeometryTests } from '../../qunit-utils';
 import {
-	ConeGeometry,
 	ConeBufferGeometry
 } from '../../../../src/geometries/ConeGeometry';
 
 export default QUnit.module( 'Geometries', () => {
-
-	QUnit.module( 'ConeGeometry', ( hooks ) => {
-
-		var geometries = undefined;
-		hooks.beforeEach( function () {
-
-			geometries = [
-				new ConeGeometry()
-			];
-
-		} );
-
-		// INHERITANCE
-		QUnit.todo( "Extending", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// INSTANCING
-		QUnit.todo( "Instancing", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// OTHERS
-		QUnit.test( 'Standard geometry tests', ( assert ) => {
-
-			runStdGeometryTests( assert, geometries );
-
-		} );
-
-	} );
 
 	QUnit.module( 'ConeBufferGeometry', ( hooks ) => {
