+++ test/unit/src/geometries/CylinderGeometry.tests.js
@@ -6,64 +6,10 @@
 
 import { runStdGeometryTests } from '../../qunit-utils';
 import {
-	CylinderGeometry,
 	CylinderBufferGeometry
 } from '../../../../src/geometries/CylinderGeometry';
 
 export default QUnit.module( 'Geometries', () => {
-
-	QUnit.module( 'CylinderGeometry', ( hooks ) => {
-
-		var geometries = undefined;
-		hooks.beforeEach( function () {
-
-			const parameters = {
-				radiusTop: 10,
-				radiusBottom: 20,
-				height: 30,
-				radialSegments: 20,
-				heightSegments: 30,
-				openEnded: true,
-				thetaStart: 0.1,
-				thetaLength: 2.0,
-			};
-
-			geometries = [
-				new CylinderGeometry(),
-				new CylinderGeometry( parameters.radiusTop ),
-				new CylinderGeometry( parameters.radiusTop, parameters.radiusBottom ),
-				new CylinderGeometry( parameters.radiusTop, parameters.radiusBottom, parameters.height ),
-				new CylinderGeometry( parameters.radiusTop, parameters.radiusBottom, parameters.height, parameters.radialSegments ),
-				new CylinderGeometry( parameters.radiusTop, parameters.radiusBottom, parameters.height, parameters.radialSegments, parameters.heightSegments ),
-				new CylinderGeometry( parameters.radiusTop, parameters.radiusBottom, parameters.height, parameters.radialSegments, parameters.heightSegments, parameters.openEnded ),
-				new CylinderGeometry( parameters.radiusTop, parameters.radiusBottom, parameters.height, parameters.radialSegments, parameters.heightSegments, parameters.openEnded, parameters.thetaStart ),
-				new CylinderGeometry( parameters.radiusTop, parameters.radiusBottom, parameters.height, parameters.radialSegments, parameters.heightSegments, parameters.openEnded, parameters.thetaStart, parameters.thetaLength ),
-			];
-
-		} );
-
-		// INHERITANCE
-		QUnit.todo( "Extending", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// INSTANCING
-		QUnit.todo( "Instancing", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// OTHERS
-		QUnit.test( 'Standard geometry tests', ( assert ) => {
-
-			runStdGeometryTests( assert, geometries );
-
-		} );
-
-	} );
 
 	QUnit.module( 'CylinderBufferGeometry', ( hooks ) => {
