+++ test/unit/src/geometries/ExtrudeGeometry.tests.js
@@ -3,27 +3,9 @@
  */
 /* global QUnit */
 
-import { ExtrudeGeometry, ExtrudeBufferGeometry } from '../../../../src/geometries/ExtrudeGeometry';
+import { ExtrudeBufferGeometry } from '../../../../src/geometries/ExtrudeGeometry';
 
 export default QUnit.module( 'Geometries', () => {
-
-	QUnit.module( 'ExtrudeGeometry', () => {
-
-		// INHERITANCE
-		QUnit.todo( "Extending", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// INSTANCING
-		QUnit.todo( "Instancing", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-	} );
 
 	QUnit.module( 'ExtrudeBufferGeometry', () => {
