+++ test/unit/src/geometries/TextGeometry.tests.js
@@ -4,45 +4,10 @@
 /* global QUnit */
 
 import {
-	TextGeometry,
 	TextBufferGeometry
 } from '../../../../src/geometries/TextGeometry';
 
 export default QUnit.module( 'Geometries', () => {
-
-	QUnit.module( 'TextGeometry', ( hooks ) => {
-
-		var geometries = undefined;
-		hooks.beforeEach( function () {
-
-			geometries = [
-				new TextGeometry()
-			];
-
-		} );
-
-		// INHERITANCE
-		QUnit.todo( "Extending", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// INSTANCING
-		QUnit.todo( "Instancing", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// OTHERS
-		QUnit.todo( 'Standard geometry tests', ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-	} );
 
 	QUnit.module( 'TextBufferGeometry', ( hooks ) => {
