+++ test/unit/src/geometries/TorusGeometry.tests.js
@@ -6,58 +6,10 @@
 
 import { runStdGeometryTests } from '../../qunit-utils';
 import {
-	TorusGeometry,
 	TorusBufferGeometry
 } from '../../../../src/geometries/TorusGeometry';
 
 export default QUnit.module( 'Geometries', () => {
-
-	QUnit.module( 'TorusGeometry', ( hooks ) => {
-
-		var geometries = undefined;
-		hooks.beforeEach( function () {
-
-			const parameters = {
-				radius: 10,
-				tube: 20,
-				radialSegments: 30,
-				tubularSegments: 10,
-				arc: 2.0,
-			};
-
-			geometries = [
-				new TorusGeometry(),
-				new TorusGeometry( parameters.radius ),
-				new TorusGeometry( parameters.radius, parameters.tube ),
-				new TorusGeometry( parameters.radius, parameters.tube, parameters.radialSegments ),
-				new TorusGeometry( parameters.radius, parameters.tube, parameters.radialSegments, parameters.tubularSegments ),
-				new TorusGeometry( parameters.radius, parameters.tube, parameters.radialSegments, parameters.tubularSegments, parameters.arc ),
-			];
-
-		} );
-
-		// INHERITANCE
-		QUnit.todo( "Extending", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// INSTANCING
-		QUnit.todo( "Instancing", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// OTHERS
-		QUnit.test( 'Standard geometry tests', ( assert ) => {
-
-			runStdGeometryTests( assert, geometries );
-
-		} );
-
-	} );
 
 	QUnit.module( 'TorusBufferGeometry', ( hooks ) => {
