+++ test/unit/src/geometries/TorusKnotGeometry.tests.js
@@ -6,59 +6,10 @@
 
 import { runStdGeometryTests } from '../../qunit-utils';
 import {
-	TorusKnotGeometry,
 	TorusKnotBufferGeometry
 } from '../../../../src/geometries/TorusKnotGeometry';
 
 export default QUnit.module( 'Geometries', () => {
-
-	QUnit.module( 'SphereGeometry', ( hooks ) => {
-
-		var geometries = undefined;
-		hooks.beforeEach( function () {
-
-			const parameters = {
-				radius: 10,
-				tube: 20,
-				tubularSegments: 30,
-				radialSegments: 10,
-				p: 3,
-				q: 2
-			};
-
-			geometries = [
-				new TorusKnotGeometry(),
-				new TorusKnotGeometry( parameters.radius ),
-				new TorusKnotGeometry( parameters.radius, parameters.tube ),
-				new TorusKnotGeometry( parameters.radius, parameters.tube, parameters.tubularSegments ),
-				new TorusKnotGeometry( parameters.radius, parameters.tube, parameters.tubularSegments, parameters.radialSegments ),
-				new TorusKnotGeometry( parameters.radius, parameters.tube, parameters.tubularSegments, parameters.radialSegments, parameters.p, parameters.q ),
-			];
-
-		} );
-
-		// INHERITANCE
-		QUnit.todo( "Extending", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// INSTANCING
-		QUnit.todo( "Instancing", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// OTHERS
-		QUnit.test( 'Standard geometry tests', ( assert ) => {
-
-			runStdGeometryTests( assert, geometries );
-
-		} );
-
-	} );
 
 	QUnit.module( 'TorusKnotBufferGeometry', ( hooks ) => {
